+++ server/src/routes/api/auth/controllers/index.ts
@@ -1,0 +1,2 @@
+export { authCallback } from './authCallback';
+export { authRequest } from './authRequest';
